Limit parallel test method count for heavy tests
For parallel tests we use Runtime#availableProcessors parallel test
method runners. For most of the tests, it's a reasonable default and it
works well, but for heavily multi-threaded tests we may overload the
system, especially if there are multiple builds running in parallel.

This change puts a more tight limit on BasicMapTest's and
AbstractNearCacheBasicTest's parallel test method count.

Potentially fixes:
https://github.com/hazelcast/hazelcast-enterprise/issues/2297
https://github.com/hazelcast/hazelcast-enterprise/issues/2481