Use submitToKeys to merge to IMap (#1107)
IMap.executeOnKeys has an issue that a copy of the EntryProcessor (EP)
is sent to each member and, additionally, one copy is sent again for
backup. The EP used for `mapWithMerging` contained a HashMap with a
value for each key. The problem was, that the map got copied many times,
wasting a lot of memory.

This commit splits the HashMap into a submap, one for each partition,
and uses one `submitToKeys` call for each partition.

Fixes #970

Also fix ReadWithPartitionIteratorP_MigrationDetectionTest: the job 
didn't create execution plans before migration was over (fixes #1107)