fix handling of rest urls
When a given url doesnt exist 404 error code should be returned instead of 400.
Also if there is any error during handling 500 error code should be returned.
If there is index out of errors it means that name of data structure cannot be
extracted because of missing /, in that case 400 is returned.