Fixes handling of exceptions in all-partitions Cache message tasks
Before #13844, exceptions thrown from reduce(map) would be caught
higher up the class hierarchy and handled via handleProcessingFailure,
properly sending an exception to the client. Since #13844 these
exceptions are just logged on the member-side. This PR restores handling
of exceptions from reduce(map) so the client is properly notified of the
exception.