Allow dynamically adding WAN publisher configuration
We introduce a way for dynamically adding WAN publisher configurations
during cluster operation. This is done on the partition threads so that
the config mutation and publisher initialisation is ordered with other
map and cache mutation operations. Even so, there are some race
conditions related to removing WAN backup events which must be handled
separately.
We also expand the WAN testing framework to allow for adding detached
WAN publisher configurations (those without source or target cluster).