Start job scan after HZ init is done (#1125)
JobCoordinationService scans for jobs that aren't currently running and
runs them. Currently the scanning starts in the middle of the Hazelcast
service initialization phase, so a job may encounter a partially
initialized Hazelcast instance.

Also: centralized JetInstance creation to avoid duplication of code between production and mock implementation for testing.