Closing all connections when heartbeat timeout
This was actually addressed in following pr.
Members and clients disconnect the connection as soon as
heartbeat timeout is detected.

https://github.com/hazelcast/hazelcast/pull/12765

This pr fixes member side heartbeat monitor. It was not closing
the connections if the connection is not an owner connection.

fixes #14094

(cherry picked from commit 17e598ccd92bd5892b830fde4b55fa7828da103e)