Fix permit acquire/release mismatch in partition anti-entropy mechanism
With the fragmented migration/replication change, backup replica synchronization
is handled per-namespace. So, for each namespace synchronization, a single permit
should be acquired. Sync requests for multiple namespaces are sent
in a single batch request to avoid multiple request/response cycles.

Problem is, a single permit was being acquired for a batch request
and but after sync completed, one permit for each namespace was released.
So, in total more than acquired permits were released after each batch.

To fix that, now a batch request will acquire permits equal to number
of namespaces that batch includes.