Fixes user code deployment provider sending wrong nested class definition
When class definition for a nested class is requested, provider only sends the nested class, not the siblings. The requesting server caches this. Later if another member asks for the sibling of the cached class, the first server falsely sends the class it has. This causes the second requesting server to fail loading the class. This PR adds a check before sending the cached class to see if the cached class is the requested one.