Cleaned up AuthenticationBaseMessageTask
The prepareAuthenticatedClientMessage had conflated concerns.
Its primary task was to create a client message because the connection
was authenticated, but a side effect is that it authenticates the
connection.

This logic has been pulled out into a seperate method and called from
the main method so that the logic of the prepareAuthenticatedClientMessage
only does what it is advertising.