Revising authentication fail status and behaviours
Client is rejected by servers when its owner is not part of the
cluster. The status of this is seperated from CREDENTIALS_FAILED.

Purpose is to change the behaviour of the client depending on the
the status.

In case of SERIALIZATION_VERSION_MISMATCH, CREDENTIALS_FAILED,
we want to directly close the client. Trying
other members on same cluster does not make sense. All members
will reject the client.

In case of OWNER_NOT_MEMBER, we want to continue trying on other
members of the same cluster. This can happen in in case of split
brain.