Adds Json preprocessing system
Json pre-processing system is a collection of utilities that makes it possible to create metadata about Json objects and query these objects faster by using the generated metadata. Metadata consists of the locations of points of interests within a JsonValue. JsonValue may be in either serialized or de-serialized form. Points of interests are:
- The offset of each scalar value from the beginning of the JsonValue
- The offset of each attribute name within the JsonValue

JsonPattern is a data structure representing logical location of a scalar value within a JsonValue. If we have a Json value A `{ "person": { "name": "abc", "age": 22 } }`. The logical location of "age" is "the second attribute of A's first attribute". We can do fast queries over Json objects by combining logical location with metadata explained above.