Improve & simplify TcpIpJoiner
Improved TcpIpJoiner to use a simpler approach;
- try to connect defined members
- ask for master address
- try to join cluster if master found
- reset master address if cannot join for a period
- rinse & repeat

(This is the same approach used in MulticastJoiner and MockJoiner.)

Additionally, a member should answer a master discovery question
only when it has already joined to a cluster. Otherwise, it can
lead other joining members to a non-existing address and eventually
none of the members can join each other.