Optimize AbstractPredicate attribute extraction
AbstractPredicate needs the type of extracted attribute so it can compare the extracted attribute with given predicate value(possibly a different type). It repeats extraction procedure just to determine type by looking at the value extracted which is already available at the end of the first extraction procedure.

The second extraction is removed.