Reduced thread creation in MetricsRegistry
THe MetricsRegistry has a scheduler where publishers can register
their tasks to be executed periodically. The issue is when publishers
with a low probe level, schedule a task. This task is scheduled even
through their probes will ignored by the metrics registry.

This is especially an issue for ephemeral clients where thread creation
matters because before this PR they would always get 2 metricsregistry
scheduler threads and with this PR a client with mandatry probe level
(default) will not create any metrics registry scheduler threads.

The thread creation in the scheduler is lazy; so as long as no tasks
get scheduled, no threads get created.