Change semantics and name of assumeThatJDK8
The primary intention of this change is to ignore W3cDomTest on JDK6.
Since we test our builds with JDK8+ versions as well, assumeThatJDK8
ignores the affected tests on JDK8+ builds, therefore changed the name
and the semantics of the assumeThatJDK8 method. From now it ignores the
affected tests on JDK6 and JDK7 only.

Fixes #14379