Fixing initial membership event when client start async
We were firing two events mistakenly when a client starts async.

The first event was empty. This pr adds a check for the first event.
If the member list is empty, it means that the initial membership
event did not arrive yet. We don't need to fire an event.

The actual event will come later and delivered to user-listener.

fixes #14364

(cherry picked from commit 83ebe2588386cb28e6c9219f611eb689c53f9368)