Add mapUsingContextAsync family of methods (#1193)
Also includes:
- add `lastForwardedWm` metric for processor

Breaking changes:

- `Processor.tryProcessWatermark` now should forward the watermark it 
receives, it's not forwarded automatically

- default implementation of `Processor.tryProcessWatermark` now throws 
an exception instead of just returning `true`. However, 
`AbstractProcessor.tryProcessWatermark` now forwards the watermark.

- `Processor.saveToSnapshot`'s contract was relaxed: Jet ensures it's 
not called while there's an unfinished item in `complete` (this is a 
compatible change, but allows removing of some code)

- `Processor.process` default implementation now throws an exception 
instead of doing nothing. It is in line with `AbstractProcessor` and 
leaving it empty is never a useful option

- `stage.mapUsingImap()` was changed to `mapUsingImapAsync()`. The 
non-partitioned version now also has to return `CompletableFuture<R>` 
instead of `R` directly.