Ignore join request with an existing member's uuid, instead of failing
When a member is restarted with the same UUID, using Hot Restart,
cluster may not notice that early and master will reject the join request.
Because it assumes that UUID is still used by an active member.

Instead of sending failure to joining member, master will ignore the request.
Terminating that member was too aggressive.