Fix cooperative emission in unordered-async-mapP (#1256)
This scenario was possible:

- `tryProcessWatermark` was called. `watermarkCounts` was empty, so it
tried to forward the watermark right away. `tryEmit` failed, so
`tryProcessWatermark` it returned `false`.

- `tryProcessWatermark` was called again. This time, the `lastReceivedWm`
was equal to the watermark, so it returned `true` without trying to
finish the watermark. This is the root of the problem.

- then `tryProcess()` was called, but `tryFlushQueue` did nothing
(because there are no in-flight items) and `tryProcess()` returned
`false` because there's an unfinished item in the outbox. The job got
stuck calling `tryProcess()` forever.

The fix is to ensure we correctly retry in `tryProcessWatermark`.

Fixes #1254