Decrease allocation rate (#1249)
* Use COWArrayList.forEach instead of iterator

This is to avoid COWIterator allocation on each iteration.
Apparently Escape Analysis does NOT eliminate the allocation. I also
tried to refactor the `run()` inside `CooperativeWorker` - in a hope
EA would kick in, but it didnt do the trick. Now with this JVM anyway:
```
openjdk version "11" 2018-09-25
OpenJDK Runtime Environment 18.9 (build 11+28)
OpenJDK 64-Bit Server VM 18.9 (build 11+28, mixed mode)
```

* Use specialized map to avoid boxing

The specialized map eliminates `frameTs` boxing in `tryProcess()`.
This is called for every item processed so it was quite significant.

* Avoid lambda instantiation on each `traverse()` invocation

According to JFR this was a frequent source of allocation. Storing
the lambda in a field avoid this issue.