Fast Json Querying with Metadata (#14393)
Adds Metadata support for records

This commit adds metadata support to Map records. Also, it adds ability to create and use metadata for AbstractJsonGetter.
AbstractRecordStore creates metadata for Json records when a record is created or updated.
This new metadata is stored in Record and is passed to querying system. Querying system passes this metadata to appropriate Getter. It is Getter's responsibility to use this metadata.