Ignore sendPartitionRuntimeState calls if node is not running
If the node is shutting down, the method can still be called. This may
cause assertion failures since partitionStateManager.isInitialized() is
false and the node is no longer master. This prints out assertion errors
in test logs even though the test passes and may mislead if the test
fails because of a different cause.
If the node is not running, we ignore the call.