Add hazelcast-full-example.yaml
The change adds hazelcast-full-example.yaml.
Besides adding this file, hazelcast-full-example.xml is slightly changed
to make it a configuration that doesn't violate any checks in the
config builders and adds a few missing configuration elements.

A test is added that verifies that the xml and yaml example files are
resolved to the same Config structure.

Copying the yaml files to the target to be part of the built zip.

(cherry picked from commit f02c9c1fa87b0b62d5ccd2e20d39f9a0ec2fde87)