Set connection type on MEMBER connections on reading protocol bytes
The connection type was previously only set in MemberEncoder upon
writing the 3 protocol bytes in response to the original connection.
This way a Bind message that arrived before the bytes were written
would be processed by BindHandler with NONE connection type -> the
connection would fail to register with the correct endpoint. This
issue only affects advanced networking configuration.