Fix handling of extracting non-terminal values from json
AbstractJsonGetter returned `null` for both:
- Non-existing paths
- Existing but non-terminal paths

When a query path points to an object or array inside a Json string, AbstractJsonGetter returned `null` for indicating this path is not suitable for querying. Only terminal values can be queried. So, we get wrong results when we run Equal or NotEqual predicate over such non-terminal path. Now we introduce a placeholder object for such paths.