Notify WAN invocations of closed connections
To speed up failover of WAN batch invocations to the target cluster, we
need to react to closed connections. In a single cluster, invocations
are cleaned up as a result of a member leave event but for WAN, we don't
receive events from the target cluster so a closed connection is the
best way of determining that the invocation has failed and can be
retried. Otherwise, the connection will hang for the duration of the
operation timeout before failing, the connection being closed and
the WAN invocation being retried.

Fixes: https://github.com/hazelcast/hazelcast-enterprise/issues/2635