Fix issue in MapStoreDataLoadingContinuesWhenNodeJoins test
The two threads in the testNoDeadLockDuringJoin test case was taking
timestamps as the last step, after a synchronization point. If thread1
got descheduled between the latch countDown and the currentTimeMillis
calls, it could take a later timestamp then thread2 and the test failed.

Now the test is using the latch.await()'s condition in thread2 to
determine if thread1 finished in time or not.

Fixes #14653