Multiple YAML configuration fixes (#14644)
Use the system properties in YAML config builders by default

Properties field to be used by variable replacers in YAML config builders was null, led to NPE if YamlConfigBuilder was building config that had variables to replace, but no properties was explicitly passed. Now it takes the system properties by default just like XML builders do.


Add missing ClientClasspathYamlConfig

The class was missing, added with tests.


Allow multiple cache-entry-listeners to be configured

Previously only listeners named cache-entry-listener in the listeners mapping could be configured under cache, which allowed only one listener configuration. Instead of mapping, listeners are configured with sequences just like in other cases. The hazelcast-fullconfig.yaml covers this.


Fix NPE if map's max-size value is not defined

If there was a map defined with max-size, but without the max-size's value an NPE was thrown. Now it reports this case as invalid configuration the same way XML configuration does.


Merge non-mapping nodes when importing

Importing YAML documents handled well only mappings. If the merging logic had to merge leaf nodes (scalars), an exception was thrown saying that the node is not a mapping. Now scalars are allowed to be merged, but only if their values are equal, otherwise an InvalidConfigurationException is thrown saying the configuration is ambiguous, including the given node's path from the root in the message.

Additionally, if the type of the nodes in the source and the target YAML is different (such as scalar vs mapping or sequence vs scalar) an InvalidConfigurationException is thrown because of ambiguous configuration, naming the different types at the node's path as the cause.

Merging sequences was not handled either, now they're merging. This allows duplicates at YAML level, because the sequences are backed by lists.