Fix a membership change bug in the Raft core
When a new member is added to a CP group, it can initialize its local
RaftNode after the new group member list is appended on other nodes.
It does not have to wait until the new member list is committed.
Dually, if the new member list is not committed and reverted, this
RaftNode must step down.