Change membership change order in the CP Subsystem
When a CP member is being removed from the CP Subsystem, it is removed
from all CP groups. For each CP group, if another CP member is going to
be added to this group as a substitute of the leaving member, the member
addition is applied before the member removal. However, member addition
can cause the majority to be increased unnecessarily. There is no need
to increase the majority here because there might be crashed members
already. To avoid this situation, we first apply the member removal, and
then the member addition.