Adds probes for bytes/packets send/received on member/client instance level
This information is availabe, but only per connection and only if we lower
the metrics level.

This PR changes that so that it is at least available on member/client instance
level (so we can see the total number of packets send/recieved and bytes send/received)
for the whole instance. This makes it easier to terminate how much bandwidth
we are actually using.