Set socket connection timeout to 10s down from infinite
Set `hazelcast.socket.connect.timeout.seconds` to 10 seconds. It was 0
before which means a connection attempt will wait infinitely until
connection succeeds or a fails with an IO error.

Reason of this change is, in some cases while restarting a pod on Kubernetes,
we found that connection attempts to former pod takes too much time
until it fails with an IO error.

Additionally, usage of `SO_LINGER` socket option in Hazelcast was wrong.
Its default value was `0` but it was used as if it's disabled. It was enabled
only when a positive value is given. But in fact, enabling `SO_LINGER` with 0
timeout is a valid option too. See https://docs.oracle.com/javase/8/docs/technotes/guides/net/articles/connection_release.html
for more info.