Fix MemberToMemberDiscoveryTest
If the two nodes configured in this test can't form a cluster initially
due to hiccups or any multicasting problem the two nodes form two
single-member cluster, then the test waits for the configured eventual
timeout and then it fails, because no merging attempt is made. This is
because default hazelcast.merge.first.run.delay.seconds is 300 seconds,
which is greater then the timeout 120 seconds. This commit decreases
this setting to 5 seconds and also decreases
hazelcast.merge.next.run.delay.seconds to 5 seconds from 120 seconds
giving the two nodes a chance to discover each other even if there is a
hiccup during the initial join.

Fixes #13942