Prevent registering a closed connection when processing bind response (#14772)
In some cases, the bind response is processed after the connections map
is cleared during split brain process. The connection is closed at that
point but it was added to the connections map regardless. Further
connections to the same address fail because there is already a
connection in the map. The commit prevents adding closed connections
into the map.

Fixes: https://github.com/hazelcast/hazelcast/issues/14737