Adds a check on the send/receive buffer configuration
The value that is set on the socket is just a hint and
the OS isn't obligated to set the values configured.

This can lead to suboptimal performance. E.g. when the
socket buffer is too small, the system will not be able
to utilize the full network bandwidth (see BDP for more
info).