Don't apply response coming from old connection
https://github.com/hazelcast/hazelcast/pull/14730
In the fix below, we changed the fetching partitions so that we get
from owner connection only. This fix checks the connection that response
comes from to see if should be applied or not. If connection is
changed, response should not be applied in order not to mess with
partition state.
We saw the problem in following issue:
related to https://github.com/hazelcast/hazelcast-enterprise/issues/2854

(cherry picked from commit c0cd00125969f4b977bfab5336abee1e5ea44945)