Fixes misinterpreting failover request failure
We were skipping retrying same cluster when a failover request failure.
With this fix, we will differentiate the cases in more detail

Failover request can fail when a connection problem occurs:
In this case, authentication will fail, but we will try new connections
on same cluster.

Failover request can fail when server is old:
In this case, authentication will fail, and we should not retry same
cluster.

Failover request can fail when server is open source:
In this case, client will get OperationNotSupportedException as cause,
and we should not retry same cluster.

fixes https://github.com/hazelcast/hazelcast-enterprise/issues/2820