Fix for partition table update on the client
Client partition table can be updated from 4 different paths
1. When partition table is not create yet and user tries to access it
for the first time
2. 10 second periodic updates
3. When member list changes
4. Via listener for server version 3.9 and after

We realized that we are applying the partition table from old connection
and messing the partition version. To keep it consistent, it should
be updated from only one connection.
https://github.com/hazelcast/hazelcast/pull/14792 was fixing only
first item. With this pr, all paths are checked for version
and connection.

fixes https://github.com/hazelcast/hazelcast-enterprise/issues/2854