Networking restart cleanup
Networking.start has been renamed to restart to make it obvious
that it can be used for restarting an already shutdown Networking
instance.

Added a flag to prevent starting an already started NioNetworking instance.

Added documentation to the NioNetworking about its abilities to deal
with restarting.

The publish all task (source of OOME) now cancels the scheduled future
when the networking is restarted. So instead of keeping the same task
registered, we unschedule the task and then reschedule it when networking
restarts. So now this unscheduling/rescheduling is in line with the
other components that get shutdown/restarted.