Upgrade checkstyle version (#14881)
Upgrade checkstyle version to the latest 8.19, to enable working properly with Java8 and using checkstyle IDE plugin. Since the metrics and checks got improved since the version we've been using until now, some modifications had to be done to the sources as well.

In terms of the number of touched classes, the most significant change is removing redundant modifiers. This change includes mostly removing public modifiers from the constructors of private internal classes, and other classes with non-public visibility. This is done as the modifiers are simply removed instead of repeating the inherited visibility.